requirements.txt
fastapi
uvicorn
pyjwt
aioredis
redis
sqlalchemy
asyncpg
alembic
psycopg2-binary
pytest
httpx
pytest-asyncio